plugins/mcp-server/requirements.txt
mcp
httpx
python-dotenv
boto3
